Slight change to function signatures

+++ examples/state.cpp
@@ -1,12 +1,12 @@
-#include "faasm/faasm.h"
+#include <faasm/faasm.h>
 
 namespace faasm {
     int exec(FaasmMemory *memory) {
         const char *key = "state_example";
 
         // Check initial state size (zero when new state is created)
-        size_t oldStateSize = memory->getStateSize(key);
-        size_t newStateSize = oldStateSize + 1;
+        long oldStateSize = memory->getStateSize(key);
+        long newStateSize = oldStateSize + 1;
 
         // Create array with one extra element
         uint8_t newState[newStateSize];
